test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

# Hacking already pins down pep8, pyflakes and flake8
hacking>=2.0,<7.1.0 # Apache-2.0
coverage>=5.2.1 # Apache-2.0
pytest>=4.6.11  # MIT
pytest-cov>=2.12.1  # MIT
stestr>=2.0.0 # Apache-2.0
python-swiftclient>=3.2.0
python-keystoneclient>=3.19.0 # Apache-2.0
boto>=2.32.1
boto3>=1.9
botocore>=1.12
requests-mock>=1.2.0 # Apache-2.0
keystonemiddleware>=4.17.0 # Apache-2.0

# Security checks
bandit>=1.1.0 # Apache-2.0

docutils>=0.15  # OSI-Approved Open Source, Public Domain
